+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.example.myapplication"
+        applicationId "com.example.selectshare"
         minSdk 21
         targetSdk 31
         versionCode 1
