+++ app/src/main/res/drawable/ic_launcher_foreground.xml
@@ -1,0 +1,17 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="108dp"
+    android:height="108dp"
+    android:viewportWidth="108"
+    android:viewportHeight="108">
+  <group android:scaleX="0.12890625"
+      android:scaleY="0.12890625"
+      android:translateX="21"
+      android:translateY="21">
+    <path
+        android:pathData="M256,256m-256,0a256,256 0,1 1,512 0a256,256 0,1 1,-512 0"
+        android:fillColor="#6C63FF"/>
+    <path
+        android:pathData="M345.5,316.514C334.227,316.514 324.14,320.957 316.427,327.917L210.665,266.459C211.407,263.053 212,259.647 212,256.092C212,252.538 211.407,249.132 210.665,245.726L315.24,184.86C323.25,192.265 333.782,196.855 345.5,196.855C370.123,196.855 390,177.011 390,152.428C390,127.844 370.123,108 345.5,108C320.877,108 301,127.844 301,152.428C301,155.982 301.593,159.388 302.335,162.794L197.76,223.66C189.75,216.256 179.218,211.665 167.5,211.665C142.877,211.665 123,231.509 123,256.092C123,280.676 142.877,300.52 167.5,300.52C179.218,300.52 189.75,295.929 197.76,288.525L303.373,350.131C302.632,353.241 302.187,356.499 302.187,359.757C302.187,383.6 321.618,403 345.5,403C369.382,403 388.813,383.6 388.813,359.757C388.813,335.914 369.382,316.514 345.5,316.514Z"
+        android:fillColor="#ffffff"/>
+  </group>
+</vector>
