+++ app/src/main/res/layout/activity_main.xml
@@ -11,18 +11,33 @@
         android:id="@+id/image_view"
         android:layout_width="match_parent"
         android:layout_height="0dp"
-        android:contentDescription="Photo"
+        android:contentDescription="@string/photo"
+        android:src="@drawable/ic_starting_photo"
         app:layout_constraintBottom_toTopOf="@id/select_photo"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toBottomOf="@id/share_image_button" />
+
+    <ImageButton
+        android:id="@+id/share_image_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:backgroundTint="@color/purple_500"
+        android:contentDescription="@string/share"
+        android:src="@drawable/ic_share_white_48dp"
+        android:visibility="invisible"
+        app:layout_constraintBottom_toTopOf="@id/image_view"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:visibility="visible" />
+
 
     <Button
         android:id="@+id/select_photo"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginVertical="12dp"
-        android:text="Select Photo"
+        android:text="@string/select_photo"
         android:textSize="24sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
